rename `compiler` module to `codegen`
