+++ qtgql/codegen/py/compiler.py
@@ -1,8 +1,8 @@
 from jinja2 import Environment, PackageLoader, select_autoescape
 
-from qtgql.compiler.objecttype import GqlType
+from qtgql.codegen.objecttype import GqlType
 
-env = Environment(loader=PackageLoader("qtgql.compiler.py"), autoescape=select_autoescape())
+env = Environment(loader=PackageLoader("qtgql.codegen.py"), autoescape=select_autoescape())
 
 SchemaTemplate = env.get_template("schema.jinja.py")
