+++ tests/test_sample_ui_itemsystem/qml/icons.py
